Crate rustful_macros

This crate provides some helpful macros for rustful, including insert_routes! and content_type!.

insert_routes!

The insert_routes! macro generates routes from the provided handlers and routing tree and adds them to the provided router. The router is then returned.

This can be useful to lower the risk of typing errors, among other things.

Example 1

#![feature(plugin)]
#[plugin]
#[no_link]
extern crate rustful_macros;

extern crate rustful;

...

let router = insert_routes!{
    TreeRouter::new(): {
        "/about" => Get: about_us,
        "/user/:user" => Get: show_user,
        "/product/:name" => Get: show_product,
        "/*" => Get: show_error,
        "/" => Get: show_welcome
    }
};

Example 2

#![feature(plugin)]
#[plugin]
#[no_link]
extern crate rustful_macros;

extern crate rustful;

...

let mut router = TreeRouter::new();
insert_routes!{
    &mut router: {
        "/" => Get: show_home,
        "home" => Get: show_home,
        "user/:username" => {Get: show_user, Post: save_user},
        "product" => {
            Get: show_all_products,

            "json" => Get: send_all_product_data
            ":id" => {
                Get: show_product,
                Post | Delete: edit_product,

                "json" => Get: send_product_data
            }
        }
    }
};

Macros

content_type!
A macro for assigning content types.

try_send!
A macro for callig send in response and aborting the handle function if it fails.